Extract buffer concerns
