+++ src/scenario/index.ts
@@ -1,3 +1,4 @@
 export * from './scenario'
 export * from './scenario-filter'
 export * from './route-timetable'
+export * from './buffer-passes'
